PR improvements
