+++ Weather App/view/SearchViewCell.swift
@@ -57,3 +57,26 @@ class SearchViewCell: UITableViewCell {
         }
     }
 }
+
+private struct Constants {
+    private init() {}
+
+    // MARK: - CityLabel's Font
+
+    static let cityLabelFont = UIFont.boldSystemFont(ofSize: 15)
+
+    // MARK: CountryLabel's Font
+
+    static let countryLabelFont = UIFont.systemFont(ofSize: 13)
+
+    // MARK: MainStack's constants
+
+    static let mainStackSpacing: CGFloat = 4
+    static let mainStackViewLeadingTrailing = 16
+    static let mainStackViewTopBottom = 10
+
+    // MARK: Cell's constants
+
+    static let cellBorderWidth: CGFloat = 2
+    static let cellCornernRadius: CGFloat = 8
+}
